updated Count code to work with new sql shiz
